migrates welcome and redeemm app router
