+++ apps/course-builder-web/src/app/(commerce)/thanks/redeem/page.tsx
@@ -1,27 +1,26 @@
 import * as React from 'react'
-import { GetServerSideProps } from 'next'
+import { notFound } from 'next/navigation'
 import { courseBuilderAdapter } from '@/db'
 import { LoginLink } from '@/path-to-purchase/post-purchase-login-link'
 
-export const getServerSideProps: GetServerSideProps = async (context) => {
-	const { query } = context
-
-	const { purchaseId } = query
-
+const ThanksRedeem = async ({
+	searchParams,
+}: {
+	searchParams: { purchaseId: string }
+}) => {
 	const { getPurchaseWithUser } = courseBuilderAdapter
 
-	const purchase = await getPurchaseWithUser(purchaseId as string)
+	const purchase = await getPurchaseWithUser(searchParams.purchaseId)
 
-	return {
-		props: {
-			email: purchase?.user?.email,
-		},
+	if (!purchase) {
+		return notFound()
 	}
-}
+	const user = await courseBuilderAdapter.getUser?.(purchase.userId as string)
 
-const ThanksRedeem: React.FC<
-	React.PropsWithChildren<{ purchase: any; email: string }>
-> = ({ email }) => {
+	if (!user) {
+		return notFound()
+	}
+
 	return (
 		<div>
 			<main className="flex flex-grow flex-col items-center justify-center px-5 pb-16 pt-5">
@@ -29,7 +28,7 @@ const ThanksRedeem: React.FC<
 					<h1 className="w-full pb-3 font-semibold uppercase tracking-wide">
 						Success!
 					</h1>
-					<LoginLink email={email} />
+					<LoginLink email={user.email} />
 				</div>
 			</main>
 		</div>
